refactor(forms): extract shared logic from `FormGroupDirective`
Ahead of the implementation of `FormArrayDirective`, extract the shared logic into an abstract class.
