+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -381,6 +381,7 @@
       "createTextNode",
       "createViewBlueprint",
       "cyclicDependencyError",
+      "declareDirectiveHostTemplate",
       "declareNoDirectiveHostTemplate",
       "decreaseElementDepthCount",
       "deepForEach",
@@ -733,6 +734,7 @@
       "stringifyForError",
       "syncViewWithBlueprint",
       "templateCreate",
+      "throwCyclicDependencyError",
       "throwError",
       "throwInvalidWriteToSignalErrorFn",
       "throwProviderNotFoundError",
