+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -11,6 +11,7 @@
       "AbstractControl",
       "AbstractControlDirective",
       "AbstractControlStatus",
+      "AbstractFormDirective",
       "AbstractFormGroupDirective",
       "AfterRenderManager",
       "AnonymousSubject",
@@ -1004,6 +1005,7 @@
       "syncPendingControls",
       "syncViewWithBlueprint",
       "templateCreate",
+      "throwCyclicDependencyError",
       "throwError",
       "throwInvalidWriteToSignalError",
       "throwInvalidWriteToSignalErrorFn",
